plugged all components types
