+++ packages/@coorpacademy-app-review/src/index.tsx
@@ -3,6 +3,7 @@ import type {} from 'redux-thunk/extend-redux'; // https://github.com/reduxjs/re
 import {AnyAction, Store} from 'redux';
 import {useSelector, useDispatch, Provider} from 'react-redux';
 import AppReviewTemplate from '@coorpacademy/components/es/template/app-review';
+import {Props as AppReviewTemplateProps} from '@coorpacademy/components/es/template/app-review/prop-types';
 
 import isEmpty from 'lodash/fp/isEmpty';
 import get from 'lodash/fp/get';
@@ -18,10 +19,10 @@ import {postProgression} from './actions/api/post-progression';
 import {mapStateToSlidesProps} from './views/slides';
 import {mapStateToSkillsProps} from './views/skills';
 
-const ConnectedApp = ({onQuitClick}: {onQuitClick: Function}): JSX.Element => {
+const ConnectedApp = ({onQuitClick}: {onQuitClick: () => void}): JSX.Element => {
   const dispatch = useDispatch();
 
-  const props = {
+  const props: AppReviewTemplateProps = {
     navigateBack: () => dispatch(navigateBack),
     viewName: useSelector(
       (state: StoreState) => state.ui.navigation[state.ui.navigation.length - 1]
