+++ packages/@coorpacademy-app-review/src/views/slides/test/index.test.ts
@@ -3,6 +3,11 @@ import identity from 'lodash/fp/identity';
 import omit from 'lodash/fp/omit';
 import pick from 'lodash/fp/pick';
 import set from 'lodash/fp/set';
+
+import {Props as PlayerProps} from '@coorpacademy/components/es/template/app-review/player/prop-types';
+import {Props as CongratsProps} from '@coorpacademy/components/es/organism/review-congrats/prop-types';
+import {Props as CongratsCardProps} from '@coorpacademy/components/es/molecule/review-card-congrats/prop-types';
+
 import {
   postProgressionResponse as createdProgression,
   postAnswerResponses,
@@ -10,7 +15,7 @@ import {
   getChoicesCorrection,
   incorrectFreeTextPostAnswerResponse
 } from '../../../test/util/services.mock';
-import {CongratsCardProps, CongratsProps, mapStateToSlidesProps} from '..';
+import {mapStateToSlidesProps} from '..';
 import type {StoreState} from '../../../reducers';
 import {freeTextSlide} from './fixtures/free-text';
 import {qcmGraphicSlide} from './fixtures/qcm-graphic';
@@ -55,6 +60,7 @@ test('should create initial props when fetched slide is not still received', t =
     closeButtonAriaLabel: 'aria-close-button',
     mode: '__revision_mode',
     skillName: '__agility',
+    hiddenSteps: false,
     steps: [
       {
         current: true,
@@ -141,11 +147,12 @@ test('should create props when first slide is on the state', t => {
     }
   };
 
-  const props = mapStateToSlidesProps(state, identity, identity);
+  const props: PlayerProps = mapStateToSlidesProps(state, identity, identity);
   t.is(props.congrats, undefined);
   t.deepEqual(omit(['onQuitClick'], props.header), {
     'aria-label': 'aria-header-wrapper',
     closeButtonAriaLabel: 'aria-close-button',
+    hiddenSteps: false,
     mode: '__revision_mode',
     skillName: '__agility',
     steps: [
@@ -249,13 +256,14 @@ test('should create props when slide is on the state and user has selected answe
     }
   };
 
-  const props = mapStateToSlidesProps(state, identity, identity);
+  const props: PlayerProps = mapStateToSlidesProps(state, identity, identity);
   t.is(props.congrats, undefined);
   t.deepEqual(omit(['onQuitClick'], props.header), {
     'aria-label': 'aria-header-wrapper',
     closeButtonAriaLabel: 'aria-close-button',
     mode: '__revision_mode',
     skillName: '__agility',
+    hiddenSteps: false,
     steps: [
       {
         current: true,
@@ -369,6 +377,7 @@ test('should verify props when first slide was answered correctly and next slide
     'aria-label': 'aria-header-wrapper',
     closeButtonAriaLabel: 'aria-close-button',
     mode: '__revision_mode',
+    hiddenSteps: false,
     skillName: '__agility',
     steps: [
       {
@@ -486,6 +495,7 @@ test('should verify props when first slide was answered with error and next slid
   t.deepEqual(omit(['onQuitClick'], props.header), {
     'aria-label': 'aria-header-wrapper',
     closeButtonAriaLabel: 'aria-close-button',
+    hiddenSteps: false,
     mode: '__revision_mode',
     skillName: '__agility',
     steps: [
@@ -556,12 +566,13 @@ test('should verify props when first slide was answered, next slide is fetched &
     }
   };
 
-  const props = mapStateToSlidesProps(state, identity, identity);
+  const props: PlayerProps = mapStateToSlidesProps(state, identity, identity);
   t.is(props.congrats, undefined);
   t.deepEqual(omit(['onQuitClick'], props.header), {
     'aria-label': 'aria-header-wrapper',
     closeButtonAriaLabel: 'aria-close-button',
     mode: '__revision_mode',
+    hiddenSteps: false,
     skillName: '__agility',
     steps: [
       {
@@ -601,7 +612,7 @@ test('should verify props when first slide was answered, next slide is fetched &
     },
     klf: undefined,
     next: {
-      ariaLabel: '_correctionNextAriaLabel',
+      'aria-label': '_correctionNextAriaLabel',
       label: '_correctionNextLabel'
     },
     type: 'right'
@@ -688,12 +699,13 @@ test('should verify props when first slide was answered incorrectly, next slide 
     }
   };
 
-  const props = mapStateToSlidesProps(state, identity, identity);
+  const props: PlayerProps = mapStateToSlidesProps(state, identity, identity);
   t.is(props.congrats, undefined);
   t.deepEqual(omit(['onQuitClick'], props.header), {
     'aria-label': 'aria-header-wrapper',
     closeButtonAriaLabel: 'aria-close-button',
     mode: '__revision_mode',
+    hiddenSteps: false,
     skillName: '__agility',
     steps: [
       {
@@ -736,7 +748,7 @@ test('should verify props when first slide was answered incorrectly, next slide 
         'To negotiate your salary when being hired, you have to establish a benchmark beforehand. In other words, you should assess the salary to which you aspire by enquiring about the remuneration paid in the same industry, the same region and the same position.'
     },
     next: {
-      ariaLabel: '_correctionNextAriaLabel',
+      'aria-label': '_correctionNextAriaLabel',
       label: '_correctionNextLabel'
     },
     type: 'wrong'
@@ -827,12 +839,13 @@ test('should verify props when currentSlideRef has changed to nextContent of pro
     }
   };
 
-  const props = mapStateToSlidesProps(state, identity, identity);
+  const props: PlayerProps = mapStateToSlidesProps(state, identity, identity);
   t.is(props.congrats, undefined);
   t.deepEqual(omit(['onQuitClick'], props.header), {
     'aria-label': 'aria-header-wrapper',
     closeButtonAriaLabel: 'aria-close-button',
     mode: '__revision_mode',
+    hiddenSteps: false,
     skillName: '__agility',
     steps: [
       {
@@ -960,13 +973,14 @@ test('should verify props when progression is in success, showing last correctio
     }
   };
 
-  const props = mapStateToSlidesProps(state, identity, identity);
+  const props: PlayerProps = mapStateToSlidesProps(state, identity, identity);
   t.is(props.congrats, undefined);
   t.deepEqual(omit(['onQuitClick'], props.header), {
     'aria-label': 'aria-header-wrapper',
     closeButtonAriaLabel: 'aria-close-button',
     mode: '__revision_mode',
     skillName: '__agility',
+    hiddenSteps: false,
     steps: [
       {
         current: false,
@@ -1004,7 +1018,7 @@ test('should verify props when progression is in success, showing last correctio
     },
     klf: undefined,
     next: {
-      ariaLabel: '_correctionNextAriaLabel',
+      'aria-label': '_correctionNextAriaLabel',
       label: '_correctionNextLabel'
     },
     resultLabel: '_right',
@@ -1082,7 +1096,7 @@ test('should verify props showing congrats', t => {
     }
   };
 
-  const props = mapStateToSlidesProps(state, identity, identity);
+  const props: PlayerProps = mapStateToSlidesProps(state, identity, identity);
   const congrats = props.congrats as CongratsProps;
   t.is(congrats.title, 'Congratulations!');
   t.is(
@@ -1214,7 +1228,7 @@ test('should verify props showing congrats, with only stars card, if user has no
     }
   };
 
-  const props = mapStateToSlidesProps(state, identity, identity);
+  const props: PlayerProps = mapStateToSlidesProps(state, identity, identity);
   const congrats = props.congrats as CongratsProps;
   t.is(congrats.title, 'Congratulations!');
   t.is(
@@ -1330,6 +1344,7 @@ test('should verify props when progression has answered a current pendingSlide',
     'aria-label': 'aria-header-wrapper',
     closeButtonAriaLabel: 'aria-close-button',
     mode: '__revision_mode',
+    hiddenSteps: false,
     skillName: '__agility',
     steps: [
       {
@@ -1437,6 +1452,7 @@ test('should verify props when progression still has a pendingSlide', t => {
     closeButtonAriaLabel: 'aria-close-button',
     mode: '__revision_mode',
     skillName: '__agility',
+    hiddenSteps: false,
     steps: [
       {
         current: false,
