+++ packages/@coorpacademy-app-review/src/views/slides/test/slide.next-slide.on-click.test.ts
@@ -1,5 +1,7 @@
 import test from 'ava';
 import identity from 'lodash/fp/identity';
+import {Props as CorrectionPopinProps} from '@coorpacademy/components/es/molecule/review-correction-popin/prop-types';
+
 import type {StoreState} from '../../../reducers';
 import {
   postAnswerResponses,
@@ -7,7 +9,7 @@ import {
   incorrectFreeTextPostAnswerResponse,
   services
 } from '../../../test/util/services.mock';
-import {CorrectionPopinProps, mapStateToSlidesProps} from '..';
+import {mapStateToSlidesProps} from '..';
 import {createTestStore} from '../../../actions/test/create-test-store';
 import {NEXT_SLIDE} from '../../../actions/ui/next-slide';
 import {freeTextSlide} from './fixtures/free-text';
