+++ packages/@coorpacademy-components/src/organism/review-congrats/test/fixtures/default.ts
@@ -1,8 +1,9 @@
 import animationLottie from '../../../../atom/lottie-wrapper/test/fixtures/confetti';
 import moleculeReviewCardStar from '../../../../molecule/review-card-congrats/test/fixtures/star';
 import moleculeReviewCardRank from '../../../../molecule/review-card-congrats/test/fixtures/rank';
+import {Props} from '../../prop-types';
 
-export const defaultProps = {
+export const defaultProps: Props = {
   'aria-label': 'Review Congratulations',
   'data-name': 'review-congrats',
   animationLottie: {...animationLottie.props, height: undefined, width: undefined},
